Burn and Mint Notice Forwarding Tags Implementation
Implementing forwarding tags to mint and burn notice allows to pass arguments to extension handlers that are listening to the mint and burn notice.

An example of this could be executing the mint action with a forwarding tag that contains a recipient you want to transfer the tokens right after minting the tokens.

Another example could be executing the burn action that allows you to do some specific tasks based on the forwarding tags you specify